feat: 修改 post 表和 post_comment 表的 dislike_count 字段的类型

+++ sql/init/tables_post.sql
@@ -21,7 +21,7 @@ CREATE TABLE post (
     content_text_struct JSONB, -- 文本扩展信息
     content_type BIT(8) NOT NULL DEFAULT 0::BIT(8), -- 0000_0000   低4位：有视频、有音频、有图片、有文本
     like_count INT NOT NULL DEFAULT 0, -- 点赞数量 
-    dislike_count INT NOT NULL DEFAULT 0, -- 异常阈值。当值达到100时，会触发人工审核。举报会提高这个数值
+    dislike_count SMALLINT NOT NULL DEFAULT 0, -- 异常阈值。当值达到100时，会触发人工审核。举报会提高这个数值
     comment_num INT  NOT NULL DEFAULT 0, -- 评论数量
     options BIT(8) NOT NULL DEFAULT 0::BIT(8), -- 0000_0000   高1~2位:  是否匿名，是否关闭评论
